+++ pages/_app.js
@@ -75,7 +75,7 @@ export default function App({ Component, pageProps }) {
   function shuffle() {
     const shuffledArray = [...cards].sort(() => Math.random() - 0.5);
     setShuffledPlayCards(shuffledArray);
-    router.push("/");
+    router.push("/play");
   }
 
   function handleActiveCards(activeCards) {
